requirements.txt
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
et-xmlfile==1.1.0
idna==3.2
iniconfig==1.1.1
numpy==1.21.2
openpyxl==3.0.8
packaging==21.0
pandas==1.3.2
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
PyPDF2==1.26.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
